Add files via upload
DB as shown on the ERD on project; unpopulated, no relationships built. 
PKs AUTO_INCREMENT; Emails and passwords NOT NULL, obvs.

In the past I've found the CHARSET / CHARACTER SET of utf8mb4 to be problematic, but if we aren't using XAMPP for hosting, I don't think it matters.

You can import with cmd if you have MySQL Server installed.